django/requirements.txt
Django~=3.1.7
django-cors-headers~=3.4.0
djangorestframework~=3.11.1
Pillow==8.1.2